+++ services/payment-service/migrations/sqls/20210722063857-init-down.sql
@@ -1,0 +1,3 @@
+/* Replace with your SQL commands */
+DROP SCHEMA IF EXISTS main CASCADE;
+