+++ services/payment-service/src/enums/status.enum.ts
@@ -1,0 +1,7 @@
+export const enum Status {
+  Paid = 'paid',
+  Refund = 'refund',
+  AlreadyPaid = 'alreadyPaid',
+  Draft = 'draft',
+  Captured = 'captured',
+}