+++ services/payment-service/src/providers/types.ts
@@ -1,0 +1,11 @@
+import {Orders} from '../models';
+import {DataObject} from '@loopback/repository';
+
+export interface IGateway {
+  create(
+    payorder: Orders,
+    paymentTemplate: string | undefined,
+  ): Promise<unknown> | DataObject<{}>;
+  charge(chargeResponse: DataObject<{}>): Promise<DataObject<{res: string}>>;
+  refund(transactionId: string): Promise<unknown> | void;
+}