+++ services/payment-service/src/repositories/transactions.repository.ts
@@ -1,0 +1,16 @@
+import {inject} from '@loopback/core';
+import {DefaultCrudRepository, juggler} from '@loopback/repository';
+import {PaymentDatasourceName} from '../keys';
+import {Transactions} from '../models';
+
+export class TransactionsRepository extends DefaultCrudRepository<
+  Transactions,
+  typeof Transactions.prototype.id
+> {
+  constructor(
+    @inject(`datasources.${PaymentDatasourceName}`)
+    dataSource: juggler.DataSource,
+  ) {
+    super(Transactions, dataSource);
+  }
+}